Updated and streamlined Block Code
- All existing block code has been updated to use the abstract class
- Animators now use triggers. Replaced coroutine code.
