+++ Assets/Prefabs/Coin Block.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 58bc8735658d0b44396b1a0d23be1fcc
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
