+++ Assets/Prefabs/Invisible Block.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4c479800ae94df14aa4d571df88d8c3a
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
