+++ Assets/Scripts/BlockScripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 953c4b37ec3058546a0a8a91357f1eba
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
