+++ Assets/Scripts/BlockScripts/Block.cs
@@ -6,14 +6,14 @@ public abstract class Block : MonoBehaviour
 {
     protected Animator anim;
 
-
     protected SpriteRenderer spriteRenderer;
     [SerializeField] protected Sprite newSprite;
 
     protected bool IsActive = true;
 
     // Start is called before the first frame update
-    protected virtual void Start()
+
+    protected virtual void Awake()
     {
         spriteRenderer = gameObject.GetComponent<SpriteRenderer>();
 
@@ -22,7 +22,7 @@ public abstract class Block : MonoBehaviour
 
     protected void OnCollisionEnter2D(Collision2D collision)
     {
-        if (IsValid() && IsActive && collision.transform.position.y < transform.position.y - 0.8)
+        if (IsValid() && IsActive && collision.transform.position.y < transform.position.y - 0.08f)
         {
             ActivateBlock();
             AnimateBlock();
@@ -46,16 +46,7 @@ public abstract class Block : MonoBehaviour
 
     protected void AnimateBlock()
     {
-        StartCoroutine(RecoilBlock());
-    }
-
-    protected IEnumerator RecoilBlock()   //makes coin block move when hit
-    {
-        anim.SetBool("recoil", true);
-
-        yield return new WaitForSeconds(0.2f);
-
-        anim.SetBool("recoil", false);
+        anim.SetTrigger("recoil");
     }
