+++ Assets/Scripts/BlockScripts/ItemBlock.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 268d581f6fb880c49961519f7fd3d7f3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
